# Switch usage from os.*File.Readdir to os.*File.ReadDir.