I suggest theses changes because I could see some improvements on my tests regarding the following benchmark:

```
import (
        "testing"

        "github.com/shirou/gopsutil/v3/net"
)

func BenchmarkConnections(b *testing.B) {
        for i := 0; i < b.N; i++ {
                net.Connections("all")
        }
}
```


```
✦ ➜ benchstat old.txt new.txt
name           old time/op    new time/op    delta
Connections-8    20.8ms ± 0%    16.7ms ± 0%   ~     (p=1.000 n=1+1)

name           old alloc/op   new alloc/op   delta
Connections-8    4.38MB ± 0%    4.01MB ± 0%   ~     (p=1.000 n=1+1)

name           old allocs/op  new allocs/op  delta
Connections-8     45.2k ± 0%     42.4k ± 0%   ~     (p=1.000 n=1+1)
```

Can you confirm that these improvements are relevants?