@Lomanic 2021-07-20
[This was introduced in go1.16](https://pkg.go.dev/os#ReadDir), while the [v3 package minimum version is 1.15](https://github.com/shirou/gopsutil/blob/4746baa8c71363bf5f0456d006136156e53c44be/v3/go.mod#L3).