@shirou 2021-07-23
Thank you for your contribution!
Go 1.15 will be out of support when 1.17 is released. And perhaps 1.17 will be release within a few month, I want to leave this PR open.